# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proyecta360 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sprints": {
      "post": {
        "summary": "Create Sprint",
        "operationId": "create_sprint_api_sprints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SprintIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Sprint Api Sprints Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stories": {
      "post": {
        "summary": "Create Story",
        "operationId": "create_story_api_stories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Story Api Stories Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stories/{story_id}": {
      "put": {
        "summary": "Update Story",
        "operationId": "update_story_api_stories__story_id__put",
        "parameters": [
          {
            "name": "story_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Story Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Update Story Api Stories  Story Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/tasks/{task_id}/scrum-summary": {
      "get": {
        "summary": "Task Scrum Summary",
        "operationId": "task_scrum_summary_api_projects__project_id__tasks__task_id__scrum_summary_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Task Scrum Summary Api Projects  Project Id  Tasks  Task Id  Scrum Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/scrum/linkable-tasks": {
      "get": {
        "summary": "Linkable Tasks",
        "operationId": "linkable_tasks_api_projects__project_id__scrum_linkable_tasks_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Linkable Tasks Api Projects  Project Id  Scrum Linkable Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SprintIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "goal": {
            "type": "string",
            "title": "Goal",
            "default": ""
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Planeado"
          },
          "velocity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Velocity",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name",
          "start_date",
          "end_date"
        ],
        "title": "SprintIn"
      },
      "StoryIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "sprint_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sprint Id"
          },
          "master_task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Master Task Id"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 1,
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Por hacer"
          },
          "points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Points",
            "default": 0
          },
          "assignee": {
            "type": "string",
            "title": "Assignee",
            "default": ""
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "default": "Media"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title"
        ],
        "title": "StoryIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}